+++ plugins/trivy/cmd/trivy/main.go
@@ -7,12 +7,9 @@ import (
 	"io/ioutil"
 	"os"
 	"os/exec"
-	"sort"
 	"strconv"
-	"strings"
 
 	"github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/image"
-	"github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/models"
 	"github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/util"
 	"github.com/sirupsen/logrus"
 )
@@ -45,21 +42,21 @@ func main() {
 	for _, i := range images {
 		logrus.Infof("%v - %v", i.Name, i.ID)
 	}
-	imagesToScan := getUnscannedImagesToScan(images, lastReport.Images)
+	imagesToScan := image.GetUnscannedImagesToScan(images, lastReport.Images, numberToScan)
 	unscannedCount := len(imagesToScan)
 	logrus.Infof("Found %d images that have never been scanned", unscannedCount)
-	imagesToScan = getImagesToRescan(images, *lastReport, imagesToScan)
+	imagesToScan = image.GetImagesToRescan(images, *lastReport, imagesToScan, numberToScan)
 	logrus.Infof("Will rescan %d additional images", len(imagesToScan) - unscannedCount)
 	logrus.Infof("Listing images to be scanned:")
 	for _, i := range imagesToScan {
 		logrus.Infof("%v - %v", i.Name, i.ID)
 	}
 	// Remove any images from the report that are no longer in the cluster
-	lastReport.Images = getMatchingImages(lastReport.Images, images, false)
+	lastReport.Images = image.GetMatchingImages(lastReport.Images, images, false)
 	// Remove any images from the report that we're going to re-scan now
-	lastReport.Images = getUnmatchingImages(lastReport.Images, imagesToScan, false)
+	lastReport.Images = image.GetUnmatchingImages(lastReport.Images, imagesToScan, false)
 	// Remove any recommendations from the report that no longer have a corresponding image in the cluster
-	lastReport.Images = getMatchingImages(lastReport.Images, images, true)
+	lastReport.Images = image.GetMatchingImages(lastReport.Images, images, true)
 
 	logrus.Infof("Starting image scans")
 	allReports := image.ScanImages(imagesToScan, maxConcurrentScans, extraFlags, false)
@@ -68,7 +65,7 @@ func main() {
 		logrus.Infof("Scanning recommendations")
 		recommendationsToScan := image.GetNewestVersionsToScan(ctx, allReports, imagesToScan)
 		// Remove any recommendations from the report that we're going to re-scan now
-		lastReport.Images = getUnmatchingImages(lastReport.Images, recommendationsToScan, true)
+		lastReport.Images = image.GetUnmatchingImages(lastReport.Images, recommendationsToScan, true)
 		logrus.Infof("Scanning %d recommended images", len(recommendationsToScan))
 		recommendationReport := image.ScanImages(recommendationsToScan, maxConcurrentScans, extraFlags, true)
 		logrus.Infof("Done scanning recommendations")
@@ -98,132 +95,6 @@ func setLogLevel() {
 	} else {
 		logrus.SetLevel(logrus.InfoLevel)
 	}
-}
-
-func getUnscannedImagesToScan(images []models.Image, lastReportImages []models.ImageDetailsWithRefs) []models.Image {
-	alreadyAdded := map[string]bool{}
-	imagesToScan := make([]models.Image, 0)
-	for _, img := range images {
-		imageSha := image.GetShaFromID(img.ID)
-		found := false
-		for _, report := range lastReportImages {
-			reportSha := image.GetShaFromID(report.ID)
-			if report.Name == img.Name && reportSha == imageSha {
-				found = true
-				break
-			}
-		}
-		if !found && !alreadyAdded[imageSha] {
-			imagesToScan = append(imagesToScan, img)
-			alreadyAdded[imageSha] = true
-		}
-	}
-	if len(imagesToScan) > numberToScan {
-		imagesToScan = imagesToScan[:numberToScan]
-	}
-	return imagesToScan
-}
-
-func getMatchingImages(baseImages []models.ImageDetailsWithRefs, toMatch []models.Image, isRecommendation bool) []models.ImageDetailsWithRefs {
-  return getImages(baseImages, toMatch, isRecommendation, true)
-}
-
-func getUnmatchingImages(baseImages []models.ImageDetailsWithRefs, toMatch []models.Image, isRecommendation bool) []models.ImageDetailsWithRefs {
-  return getImages(baseImages, toMatch, isRecommendation, false)
-}
-
-func getImages(baseImages []models.ImageDetailsWithRefs, toMatch []models.Image, isRecommendation bool, match bool) []models.ImageDetailsWithRefs {
-	filtered := make([]models.ImageDetailsWithRefs, 0)
-	isMatch := convertImagesToMap(toMatch)
-	isRepoMatch:= imagesRepositoryMap(toMatch)
-	for _, im := range baseImages {
-		if !isRecommendation {
-			imageSha := image.GetShaFromID(im.ID)
-			if im.RecommendationOnly || isMatch[imageSha] == match {
-				filtered = append(filtered, im)
-			}
-		} else {
-			// For recommendations, we match only on repo name, not on full image ID
-			parts := strings.Split(im.Name, ":")
-			key := image.GetRecommendationKey(parts[0], image.GetSpecificToken(parts[1]))
-			if !im.RecommendationOnly || isRepoMatch[key] == match{
-				filtered = append(filtered, im)
-			}
-		}
-	}
-	return filtered
-}
-
-
-func getImagesToRescan(images []models.Image, lastReport models.MinimizedReport, imagesToScan []models.Image) []models.Image {
-	sort.Slice(lastReport.Images, func(a, b int) bool {
-		return lastReport.Images[a].LastScan == nil || lastReport.Images[b].LastScan != nil && lastReport.Images[a].LastScan.Before(*lastReport.Images[b].LastScan)
-	})
-	for _, report := range lastReport.Images {
-		reportSha := image.GetShaFromID(report.ID)
-		if !report.RecommendationOnly {
-			for _, img := range images {
-				imageSha := image.GetShaFromID(img.ID)
-				if report.Name == img.Name && reportSha == imageSha {
-					if len(imagesToScan) < numberToScan {
-						imagesToScan = append(imagesToScan, img)
-						break
-					} else {
-						return imagesToScan
-					}
-				}
-			}
-		}
-	}
-	return imagesToScan
-}
-
-func getRecommendationImagesToKeep(images []models.Image, lastReport models.MinimizedReport, recommendationsToScan []models.Image) []models.ImageDetailsWithRefs {
-	imagesToKeep := make([]models.ImageDetailsWithRefs, 0)
-	sort.Slice(lastReport.Images, func(a, b int) bool {
-		return lastReport.Images[a].LastScan == nil || lastReport.Images[b].LastScan != nil && lastReport.Images[a].LastScan.Before(*lastReport.Images[b].LastScan)
-	})
-	newRecommendations := convertImagesToMap(recommendationsToScan)
-	clusterImagesMap := imagesRepositoryMap(images)
-	for _, report := range lastReport.Images {
-		reportSha := image.GetShaFromID(report.ID)
-		// We must keep images recommendations for those still in the cluster but not scanned at this time
-		if report.RecommendationOnly {
-			parts := strings.Split(report.Name, ":")
-			if len(parts) == 2 {
-				key := image.GetRecommendationKey(parts[0], image.GetSpecificToken(parts[1]))
-				// Add old recommendations only if we have the images they are for still running in the cluster
-				if _, found := clusterImagesMap[key]; found {
-					// Add old recommendations only if we have not scanned for new recommendations
-					if _, found := newRecommendations[reportSha]; !found {
-						imagesToKeep = append(imagesToKeep, report)
-					}
-				}
-			}
-		}
-	}
-	return imagesToKeep
-}
-
-func convertImagesToMap(list []models.Image) map[string]bool {
-	m := map[string]bool{}
-	for _, img := range list {
-		sha := image.GetShaFromID(img.ID)
-		m[sha] = true
-	}
-	return m
-}
-
-func imagesRepositoryMap(list []models.Image) map[string]bool {
-	m := map[string]bool{}
-	for _, img := range list {
-		parts := strings.Split(img.Name, ":")
-		if len(parts) == 2 {
-			key := image.GetRecommendationKey(parts[0], image.GetSpecificToken(parts[1]))
-			m[key] = true
-		}
-	}
-	return m
 }
 
 func setEnv() {
